+++ community/graph-algo/src/test/java/org/neo4j/graphalgo/shortestpath/FloydWarshallTest.java
@@ -43,13 +43,13 @@ public class FloydWarshallTest extends Neo4jAlgoTestCase
     public void testMinimal()
     {
         graph.makeEdge( "a", "b", "cost", (double) 1 );
-        graph.makeEdge( "a", "c", "cost", (double) 1 );
-        graph.makeEdge( "a", "d", "cost", (double) 1 );
-        graph.makeEdge( "a", "e", "cost", (double) 1 );
+        graph.makeEdge( "a", "c", "cost", (float) 1 );
+        graph.makeEdge( "a", "d", "cost", (long) 1 );
+        graph.makeEdge( "a", "e", "cost", (int) 1 );
         graph.makeEdge( "b", "c", "cost", (double) 1 );
-        graph.makeEdge( "c", "d", "cost", (double) 1 );
-        graph.makeEdge( "d", "e", "cost", (double) 1 );
-        graph.makeEdge( "e", "b", "cost", (double) 1 );
+        graph.makeEdge( "c", "d", "cost", (byte) 1 );
+        graph.makeEdge( "d", "e", "cost", (short) 1 );
+        graph.makeEdge( "e", "b", "cost", (byte) 1 );
         FloydWarshall<Double> floydWarshall = new FloydWarshall<Double>( 0.0,
             Double.MAX_VALUE, Direction.OUTGOING,
             CommonEvaluators.doubleCostEvaluator( "cost" ),
@@ -71,10 +71,10 @@ public class FloydWarshallTest extends Neo4jAlgoTestCase
     public void testPath()
     {
         graph.makeEdge( "a", "b", "cost", (double) 1 );
-        graph.makeEdge( "b", "c", "cost", (double) 1 );
-        graph.makeEdge( "c", "d", "cost", (double) 1 );
-        graph.makeEdge( "d", "e", "cost", (double) 1 );
-        graph.makeEdge( "e", "f", "cost", (double) 1 );
+        graph.makeEdge( "b", "c", "cost", (float) 1 );
+        graph.makeEdge( "c", "d", "cost", (int) 1 );
+        graph.makeEdge( "d", "e", "cost", (long) 1 );
+        graph.makeEdge( "e", "f", "cost", (byte) 1 );
         FloydWarshall<Double> floydWarshall = new FloydWarshall<Double>( 0.0,
             Double.MAX_VALUE, Direction.OUTGOING,
             CommonEvaluators.doubleCostEvaluator( "cost" ),
